requirements.txt
# pip3 install -r requirements.txt -i https://pypi.douban.com/simple

DBUtils>=1.3
Flask>=1.0.2
PyYAML>=5.1.2
sanic>=19.12.2
PyMySQL>=0.9.3
pykafka>=2.8.0
pymongo>=3.10.1
paramiko>=2.6.0
influxdb>=5.2.3
aliyun-python-sdk-core>=2.13.13
